# api-definitions/Repository.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Repository API",
        "version": "v1"
    },
    "paths": {
...
        "/repository/user-profile/{userProfileId}": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "userProfileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/user-profile/by-identity-id/{identityId}": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "identityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/user-profile/by-xtremeidiots-id/{xtremeidiotsId}": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "xtremeIdiotsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/user-profile": {
            "get": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "filterString",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/UserProfilesOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "UserProfile"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "UserProfile"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/user-profile/{userProfileId}/claims": {
            "patch": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "userProfileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "userProfileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/repository/user-profile/{userProfileId}/claims/{userProfileClaimId}": {
            "delete": {
                "tags": [
                    "UserProfile"
                ],
                "parameters": [
                    {
                        "name": "userProfileId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "userProfileClaimId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserProfilesOrder": {
                "enum": [
                    "DisplayNameAsc",
                    "DisplayNameDesc"
                ],
                "type": "string",
                "format": "int32"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}